Tasks now use Doubles instead of Floats

+++ src/paralleltasks/GetPopulationTask.java
@@ -7,13 +7,13 @@ import java.util.concurrent.RecursiveTask;
 /*
    1) This class is the parallel version of the getPopulation() method from version 1 for use in version 2
    2) SEQUENTIAL_CUTOFF refers to the maximum number of census groups that should be processed by a single parallel task
-   3) The float parameters(w, s, e, n) represent the bounds of the query rectangle
+   3) The double parameters(w, s, e, n) represent the bounds of the query rectangle
    4) The compute method returns an Integer representing the total population contained in the query rectangle
  */
 public class GetPopulationTask extends RecursiveTask<Integer> {
     CensusGroup[] censusGroups;
     int lo, hi;
-    float w, s, e, n;
+    double w, s, e, n;
     final static int SEQUENTIAL_CUTOFF = 1000;
 
     public GetPopulationTask(CensusGroup[] censusGroups, int lo, int hi, double w, double s, double e, double n) {
